+++ .vscode/settings.json
@@ -2,9 +2,6 @@
   "search.exclude": {
     "**/.sst": true
   },
-  "editor.codeActionsOnSave": {
-    "source.organizeImports": "explicit"
-  },
   "editor.tabSize": 2,
   "editor.insertSpaces": true,
   "editor.formatOnPaste": false, // required
