+++ backend/infra/ConfigStack.ts
@@ -41,12 +41,6 @@ export function ConfigStack({ stack, app }: StackContext) {
       permissions: [[financialDataTable.cdk.table, 'grantReadWriteData']],
     }
   );
-  const applyForLoanHandler = new Function(stack, 'applyForLoanHandlerLambda', {
-    handler: 'src/adaptors/primary/api-gw-apply-for-loan.handler',
-    functionName: app.logicalPrefixedName('ApplyForLoan'),
-    bind: [FINANCIAL_DATA_TABLE_NAME],
-    permissions: [[financialDataTable.cdk.table, 'grantReadWriteData']],
-  });
 
   const createBorrowerProfileHandler = new Function(
     stack,
@@ -84,7 +78,6 @@ export function ConfigStack({ stack, app }: StackContext) {
       generateApiSpec({
         calculateBorrowingCapacityHandlerArn:
           calculateBorrowingCapacityHandler.functionArn,
-        applyForLoanHandlerArn: applyForLoanHandler.functionArn,
         createBorrowerProfileHandlerArn:
           createBorrowerProfileHandler.functionArn,
       })
@@ -116,19 +109,7 @@ export function ConfigStack({ stack, app }: StackContext) {
       },
     })
   );
-  applyForLoanHandler.grantInvoke(
-    new iam.ServicePrincipal('apigateway.amazonaws.com', {
-      conditions: {
-        ArnLike: {
-          'aws:SourceArn': api.arnForExecuteApi(
-            'POST',
-            '/loan',
-            api.deploymentStage.stageName
-          ),
-        },
-      },
-    })
-  );
+
   Config.Parameter.create(stack, {
     API_ENDPOINT: api.url,
   });
