+++ backend/src/use-cases/loan-assessment/process-loan-application.ts
@@ -14,30 +14,4 @@ export const processLoanApplication: ProcessLoanApplicationPort = async ({
   grossAnnualIncome,
   employmentStatus,
   monthlyExpenses,
-}) => {
-  const borrowerProfile = await getBorrowerProfile(borrowerEmail);
-  if (!borrowerProfile) {
-    throw new BorrowerProfileDoesNotExistError('Borrower profile not found');
-  }
-  const { creditScore, dob } = borrowerProfile;
-  const loanApplicationStatus = assessLoanApplication({
-    grossAnnualIncome,
-    employmentStatus,
-    monthlyExpenses,
-    creditScore,
-    age: calculateAge(new Date(dob)),
-  });
-
-  await putLoanApplication({
-    borrowerEmail,
-    loanApplicationId: randomUUID(),
-    timestamp: new Date().toISOString(),
-    creditScore,
-    grossAnnualIncome,
-    monthlyExpenses,
-    loanApplicationStatus,
-    employmentStatus,
-  });
-
-  return loanApplicationStatus;
-};
+}) => {};
